+++ src/include/physics/ObjectSet.h
@@ -27,12 +27,12 @@ class Object {
         Vec3* velocity;
         Vec3* gravity;
 
-        double opacity;
+        float opacity;
 
-        double mass;
-        double gravityFactor;
-        double frictionFactor;
-        double bounceFactor;
+        float mass;
+        float gravityFactor;
+        float frictionFactor;
+        float bounceFactor;
 
 
         // Constructor
@@ -53,28 +53,28 @@ class Object {
 
         // Moves the object in space by the specified distance.
         Object* move(Vec3* dist);
-        Object* move(double dx, double dy, double dz);
+        Object* move(float dx, float dy, float dz);
 
         // Scales the object by a given factor on each axis
-        Object* scaleBy(double factor);
-        Object* scaleBy(double fx, double fy, double fz);
+        Object* scaleBy(float factor);
+        Object* scaleBy(float fx, float fy, float fz);
 
         // Rotates the object by the specified angles. rotates around (0, 0, 0) if no around vector is given
         Object* rotate(Vec3* angle, Vec3* around);
-        Object* rotate(double yaw, double pitch, double roll, Vec3* around /* default value = nullptr */);
+        Object* rotate(float yaw, float pitch, float roll, Vec3* around /* default value = nullptr */);
 
         // Rotates the object around its center
         Object* rotateSelf(Vec3* angle);
-        Object* rotateSelf(double yaw, double pitch, double roll);
+        Object* rotateSelf(float yaw, float pitch, float roll);
 
         // Sets the color of the object
         Object* setColor(uint32 color);
 
         // Updates all the physics variables of this object based on a delta time
-        void doPhysics(double dt);
+        void doPhysics(float dt);
 
         // Simple collision check. the object will act as though there is an infinite mass plane at a y level
-        void doFloorCollision(double y);
+        void doFloorCollision(float y);
 
         // Updates the mesh if any vectors have been changed
         void update();
@@ -153,30 +153,30 @@ class ObjectSet {
 
         // Changes the position of all the objects
         void moveAll(Vec3* dist);
-        void moveAll(double dx, double dy, double dz);
+        void moveAll(float dx, float dy, float dz);
 
         // Sets the position of all the objects
         void setPosAll(Vec3* pos);
-        void setPosAll(double x, double y, double z);
+        void setPosAll(float x, float y, float z);
 
         // Adds velocity to all the objects
         void addVelocityAll(Vec3* v);
-        void addVelocityAll(double vx, double vy, double vz);
+        void addVelocityAll(float vx, float vy, float vz);
 
         // Sets the velocity of all the objects
         void setVelocityAll(Vec3* v);
-        void setVelocityAll(double vx, double vy, double vz);
+        void setVelocityAll(float vx, float vy, float vz);
 
         // Sets the gravitational acceleration for all objects in the set
         void setGravityAll(Vec3* gravity);  // This copies the values from thi Vec3 into instance variables, this one needs to be handled accordingly
-        void setGravityAll(double gx, double gy, double gz);  // Sets all the instance variables to these values
-        void setGravityAll(double gy);  // Sets the gravity to just down or up
+        void setGravityAll(float gx, float gy, float gz);  // Sets all the instance variables to these values
+        void setGravityAll(float gy);  // Sets the gravity to just down or up
 
         // Sets the opacity of all the objects
-        void setOpacityAll(double opacity);
+        void setOpacityAll(float opacity);
 
         // Calls Object->doPhysics for all in the set
-        void doAllPhysics(double dt);
+        void doAllPhysics(float dt);
 
         // Updates any objects where vectors have been changed
         void updateAll();
@@ -186,11 +186,11 @@ class ObjectSet {
 
         // Projects, sorts, and draws all the objects in the set, in order from furthest away to closest, optional opacity value from 0-1
         void drawAll(Drawer* drawer, Camera* camera, Display* display);
-        void drawAll(Drawer* drawer, Camera* camera, Display* display, double opacity);
+        void drawAll(Drawer* drawer, Camera* camera, Display* display, float opacity);
 
         // Does the same as drawAll(), but also draws red normal vectors on all the tris
         void drawAllWithNormals(Drawer* drawer, Camera* camera, Display* display);
-        void drawAllWithNormals(Drawer* drawer, Camera* camera, Display* display, double opacity);
+        void drawAllWithNormals(Drawer* drawer, Camera* camera, Display* display, float opacity);
 
         // Logs all of the nodes in the order they stand in the list
         inline void log() { this->list->log(); }
